feat: top 5 user recommendation method added
