+++ src/main/java/org/example/proyect/recommendation/UserMatcher.java
@@ -1,0 +1,47 @@
+package org.example.proyect.recommendation;
+
+import org.example.fuzzy_matcher.component.MatchService;
+import org.example.fuzzy_matcher.domain.*;
+import org.example.proyect.objects.User;
+
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+public class UserMatcher {
+    private MatchService matchService;
+
+    public UserMatcher(MatchService matchService) {
+        this.matchService = matchService;
+    }
+
+    public List<Document> getTopMatches(Document userDocument, List<Document> documents, int topN) {
+        Map<Document, List<Match<Document>>> matches = matchService.applyMatch(documents);
+        List<Match<Document>> userMatches = matches.get(userDocument);
+        userMatches.sort(Comparator.comparing(Match<Document>::getScore, Comparator.comparingDouble(Score::getResult)).reversed());
+        return userMatches.stream()
+                .limit(topN)
+                .map(Match::getMatchedWith)
+                .collect(Collectors.toList());
+    }
+
+    public Document getUserDocumentToMatch(List<Document> documents, String userId) {
+        return documents.stream().filter(document -> document.getKey().equals(userId)).findFirst().orElse(null);
+    }
+
+    public void printRecommendedUsers(List<Document> recommendedUsers, List<User> users) {
+        for (Document document : recommendedUsers) {
+            for (User user: users) {
+                if (document.getKey().equals(String.valueOf(user.getId()))) {
+                    System.out.println("User recommended: " + user.getName());
+                    System.out.println("User favorites genres: " + user.getFavoriteGenres());
+                    System.out.println("User favorite artists: " + user.getFavoriteArtists());
+                    System.out.println("User favorite songs: " + user.getFavoriteSongs());
+                    System.out.println("User recently played: " + user.getRecentlyPlayed());
+                }
+            }
+            System.out.println("--------------------");
+        }
+    }
+}
