+++ src/test/java/objects/user/UserDocumentTest.java
@@ -2,7 +2,7 @@ package objects.user;
 
 import org.example.fuzzy_matcher.domain.Document;
 import org.example.fuzzy_matcher.domain.Element;
-import org.example.objects.User;
+import org.example.proyect.objects.User;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
 
@@ -14,7 +14,7 @@ import static org.example.fuzzy_matcher.domain.ElementType.EMAIL;
 public class UserDocumentTest {
     @Test
     public void childCountAndUnmatchedCountTest() {
-        Document d1 = getUser1().toDocument();
+        Document d1 = getUser1().getDocument();
         System.out.println(d1);
 
         Document d2 = new Document.Builder("2")
